Refactor merging of topdDocs and scoreDocs into separate classes
